chore: 디테일한 디자인 수정, ImgUploadBtn 통일 - #11
