+++ client/src/components/dashboard/ImgUploadBtn.tsx
@@ -1,9 +1,12 @@
-import React, { useRef } from "react";
+import React, { useRef, useState } from "react";
+
+import { POST } from "@/axios/POST";
 // Assets
 import { BsCamera } from "react-icons/bs";
 
 export default function ImgUploadBtn() {
   const fileInputRef = useRef<HTMLInputElement>(null);
+  const [img, setImg] = useState([]);
 
   const handleInput = () => {
     fileInputRef.current?.click();
@@ -21,7 +24,18 @@ export default function ImgUploadBtn() {
       formData.append("photos", selectedImg[i]);
     }
 
-    POSTImg(formData); // 💡 결과 리턴
+    // 로컬 이미지 업로드 API
+    const uploadImg = async () => {
+      const data = await POST("/upload/1", formData, {
+        "Content-Type": "multipart/form-data",
+        token: 1,
+      });
+      if (data != null) {
+        setImg(data);
+      }
+    };
+
+    uploadImg();
   };
 
   return (
@@ -34,7 +48,7 @@ export default function ImgUploadBtn() {
         ref={fileInputRef}
         onChange={handleImgChange}
       />
-      <BsCamera size={40} onClick={handleInput} />
+      <BsCamera size={30} onClick={handleInput} />
     </div>
   );
 }
